WIP: Add more custom steps, streamline cosa build + kola workflow
This changes in significant ways how this shared library is set up.
There is now a much bigger emphasis on the `cosa build` workflow. For
example, there is now a `cosaPod` and which default to the cosa image.
There's a `buildroot` parameter one can use to request the buildroot
image instead (and of course, one can still specify the image they want.

There is also a `fcosBuild` custom step, which will automatically build
FCOS, optionally building and/or layering some directories, then run
`cosa kola`.